src/requirements.txt
flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
google-cloud-firestore==2.11.1